Polish code and upgrade to 1.5.1.RELEASE
