+++ jcart-admin/src/main/resources/application-default.properties
@@ -8,3 +8,6 @@ server.ssl.keyStoreType=PKCS12
 server.ssl.keyAlias=jcartadmintomcat
 
 security.filter-order=0
+
+########### Thyemeleaf ###########
+spring.thymeleaf.mode=HTML
