+++ jcart-core/src/main/java/com/sivalabs/jcart/security/SecurityService.java
@@ -1,187 +1,40 @@
 package com.sivalabs.jcart.security;
 
-import static java.util.Objects.isNull;
-import static java.util.Objects.nonNull;
-import static java.util.stream.Collectors.toList;
-
 import java.util.List;
-import java.util.UUID;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
-
-import com.sivalabs.jcart.JCartException;
 import com.sivalabs.jcart.entities.Permission;
 import com.sivalabs.jcart.entities.Role;
 import com.sivalabs.jcart.entities.User;
 
-/**
- * @author Siva
- * @author rajakolli
- */
-@Service
-@Transactional
-public class SecurityService
+public interface SecurityService
 {
-    private static final String INVALID_EMAILADDRESS = "Invalid email address";
-    private UserRepository userRepository;
-    private PermissionRepository permissionRepository;
-    private RoleRepository roleRepository;
 
-    /**
-     * Spring {@link Autowired} Constructor Injection
-     * 
-     * @param userRepository
-     * @param permissionRepository
-     * @param roleRepository
-     */
-    public SecurityService(UserRepository userRepository,
-            PermissionRepository permissionRepository, RoleRepository roleRepository)
-    {
-        super();
-        this.userRepository = userRepository;
-        this.permissionRepository = permissionRepository;
-        this.roleRepository = roleRepository;
-    }
+    User findUserByEmail(String userEmail);
 
-    public User findUserByEmail(String email)
-    {
-        return userRepository.findByEmail(email);
-    }
+    List<Permission> getAllPermissions();
 
-    public String resetPassword(String email)
-    {
-        User user = findUserByEmail(email);
-        if (isNull(user))
-        {
-            throw new JCartException(INVALID_EMAILADDRESS);
-        }
-        String uuid = UUID.randomUUID().toString();
-        user.setPasswordResetToken(uuid);
-        return uuid;
-    }
+    List<Role> getAllRoles();
 
-    public void updatePassword(String email, String token, String password)
-    {
-        User user = findUserByEmail(email);
-        if (isNull(user))
-        {
-            throw new JCartException(INVALID_EMAILADDRESS);
-        }
-        if (!StringUtils.hasText(token) || !token.equals(user.getPasswordResetToken()))
-        {
-            throw new JCartException("Invalid password reset token");
-        }
-        user.setPassword(password);
-        user.setPasswordResetToken(null);
-    }
+    Role createRole(Role role);
 
-    public boolean verifyPasswordResetToken(String email, String token)
-    {
-        User user = findUserByEmail(email);
-        if (isNull(user))
-        {
-            throw new JCartException(INVALID_EMAILADDRESS);
-        }
-        if (!StringUtils.hasText(token) || !token.equals(user.getPasswordResetToken()))
-        {
-            return false;
-        }
-        return true;
-    }
+    Role getRoleById(Integer id);
 
-    public List<Permission> getAllPermissions()
-    {
-        return permissionRepository.findAll();
-    }
+    Role updateRole(Role role);
 
-    public List<Role> getAllRoles()
-    {
-        return roleRepository.findAll();
-    }
+    String resetPassword(String email);
 
-    public Role getRoleByName(String roleName)
-    {
-        return roleRepository.findByName(roleName);
-    }
+    boolean verifyPasswordResetToken(String email, String token);
 
-    public Role createRole(Role role)
-    {
-        Role roleByName = getRoleByName(role.getName());
-        if (nonNull(roleByName))
-        {
-            throw new JCartException("Role " + role.getName() + " already exist");
-        }
-        List<Permission> persistedPermissions = role.getPermissions().stream()
-                .filter(permission -> nonNull(permission.getId()))
-                .map(permission -> permissionRepository.findOne(permission.getId()))
-                .collect(toList());
-        role.setPermissions(persistedPermissions);
-        return roleRepository.save(role);
-    }
+    void updatePassword(String email, String token, String encodedPwd);
 
-    public Role updateRole(Role role)
-    {
-        Role persistedRole = getRoleById(role.getId());
-        if (isNull(persistedRole))
-        {
-            throw new JCartException("Role " + role.getId() + " doesn't exist");
-        }
-        persistedRole.setDescription(role.getDescription());
+    List<User> getAllUsers();
 
-        List<Permission> updatedPermissions = role.getPermissions().stream()
-                .filter(permission -> nonNull(permission.getId()))
-                .map(permission -> permissionRepository.findOne(permission.getId()))
-                .collect(toList());
-        persistedRole.setPermissions(updatedPermissions);
-        return roleRepository.save(persistedRole);
-    }
+    User createUser(User user);
 
-    public Role getRoleById(Integer id)
-    {
-        return roleRepository.findOne(id);
-    }
+    User getUserById(Integer id);
 
-    public User getUserById(Integer id)
-    {
-        return userRepository.findOne(id);
-    }
+    User updateUser(User user);
 
-    public List<User> getAllUsers()
-    {
-        return userRepository.findAll();
-    }
-
-    public User createUser(User user)
-    {
-        User userByEmail = findUserByEmail(user.getEmail());
-        if (nonNull(userByEmail))
-        {
-            throw new JCartException("Email " + user.getEmail() + " already in use");
-        }
-        List<Role> persistedRoles = user.getRoles().stream()
-                .filter(role -> nonNull(role.getId()))
-                .map(role -> roleRepository.findOne(role.getId())).collect(toList());
-        user.setRoles(persistedRoles);
-
-        return userRepository.save(user);
-    }
-
-    public User updateUser(User user)
-    {
-        User persistedUser = getUserById(user.getId());
-        if (isNull(persistedUser))
-        {
-            throw new JCartException("User " + user.getId() + " doesn't exist");
-        }
-
-        List<Role> updatedRoles = user.getRoles().stream()
-                .filter(role -> nonNull(role.getId()))
-                .map(role -> roleRepository.findOne(role.getId())).collect(toList());
-        persistedUser.setRoles(updatedRoles);
-        return userRepository.save(persistedUser);
-    }
+    Role getRoleByName(String name);
 
 }
